refactor: padroniza escrita de todos os arquivos index.tsx
